+++ vitest.config.ts
@@ -28,7 +28,7 @@ export default defineConfig({
             // a red build green.
             thresholds: {
                 statements: 94,
-                branches: 90,
+                branches: 95,
                 functions: 95,
                 lines: 94,
             },